use rfc2047 crate from @valodim and remove dc_strencode.rs completely
