+++ Cargo.toml
@@ -19,6 +19,7 @@ num-derive = "0.2.5"
 num-traits = "0.2.6"
 lettre = { git = "https://github.com/deltachat/lettre", branch = "feat/mail" }
 lettre_email = { git = "https://github.com/deltachat/lettre", branch = "feat/mail" }
+rfc2047 = "0.1.1"
 async-imap = { git = "https://github.com/async-email/async-imap", branch="master" }
 async-tls = "0.6"
 async-std = { version = "1.0", features = ["unstable"] }
